Refactor SDK into shared, uid2 and euid folders
