+++ src/shared/callbackManager.ts
