+++ src/shared/cookieManager.ts
