+++ src/shared/cstgCrypto.ts
