+++ src/shared/encoding/base64.ts
