+++ src/shared/encoding/hash.ts
