+++ src/shared/initCallbacks.ts
