+++ src/shared/integrationTests/async.spec.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, jest, test } from '@jest/globals';
 
 import * as mocks from '../mocks';
-import { __uid2InternalHandleScriptLoad, sdkWindow, UID2 } from '../uid2Sdk';
+import { __uid2InternalHandleScriptLoad, sdkWindow, UID2 } from '../../uid2/uid2Sdk';
 import { EventType } from '../callbackManager';
 
 let callback: any;
