+++ src/shared/integrationTests/clientSideTokenGeneration.test.ts
@@ -2,7 +2,7 @@ import * as mocks from '../mocks';
 import { NAME_CURVE } from '../mocks';
 import { base64ToBytes, bytesToBase64 } from '../encoding/base64';
 import { EventType } from '../callbackManager';
-import { sdkWindow, UID2 } from '../uid2Sdk';
+import { sdkWindow, UID2 } from '../../uid2/uid2Sdk';
 
 let uid2: UID2;
 let xhrMock: any;
