+++ src/shared/integrationTests/compatibility.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, jest, test } from '@jest/globals';
 
 import * as mocks from '../mocks';
-import { sdkWindow, UID2 } from '../uid2Sdk';
+import { sdkWindow, UID2 } from '../../uid2/uid2Sdk';
 
 let callback: any;
 let uid2: UID2;
