+++ src/shared/integrationTests/testEnvironment.test.ts
@@ -1,6 +1,6 @@
 import { test } from '@jest/globals';
 
-import { UID2 } from '../uid2Sdk';
+import { UID2 } from '../../uid2/uid2Sdk';
 
 test('Init can be constructed', () => {
   const _uid2 = new UID2();
