+++ src/shared/localStorageManager.ts
