+++ src/shared/mocks.ts
@@ -1,6 +1,6 @@
 import * as jsdom from 'jsdom';
 import { Cookie } from 'tough-cookie';
-import { UID2 } from './uid2Sdk';
+import { UID2 } from '../uid2/uid2Sdk';
 import { Identity } from './Identity';
 import { base64ToBytes, bytesToBase64 } from './encoding/base64';
 import * as crypto from 'crypto';
