+++ src/shared/product.ts
