+++ src/shared/storageManager.ts
