+++ src/shared/unitTests/uid2CstgBox.test.ts
